Automatically start the worker pool if a job is submitted (#5039)
The worker pool is a thread pool that allows multiple jobs to run
concurrently. The worker pool is normally started only when a full node
is initialized. If a full node is not initialized, then no separate
thread is started, and jobs are not run concurrently but sequentially.

This means that the performance of CLI commands that use the worker
pool, like `wallet:rescan`, can vary drastically depending on whether a
full node is running or not. This behavior is not documented and far
from obvious from a user prospective.

To always offer the best performance, this commit changes the worker
pool to ensure that threads are always spawned before executing any job.