adds networkId to wallet, accountImport
adds networkId to wallet and wallet constructor. node wallets and standalone
wallets can only run on a single network without reset. the networkId is stored
in the 'internal' file, so it is not necessary to persist networkId in wallet
metadata

adds networkId field to 'AccountImport' type to indicate which networkId an
account was running on at export time. the network that an account was running
on is only important when importing the account to another node

uses networkId to determine if imported account birthday matches network instead
of looking up block from chain. this will avoid resetting the account birthday
if imported to a node that hasn't synced to the birthday yet