Change `DecryptNotesRequest` to require a common list of accounts and options
All the use cases for `DecryptNotesTask` involve passing a bunch of
notes to decrypt for the same account(s), and with the same options.  To
save bandwidth, and to allow future optimizations, it makes sense to
group all account keys into a single input parameter, so that they don't
have to be repeated for each and every note.

This changes the request serialization size from `O(accounts * notes)`
to `O(accounts + notes)`.