Fix transactions cmd not handling streaming errors
The error here is that the error in getAccountTransactionsStream becomes
unhandled, because we pause the event loop to run getNetworkInfo. The
error handler for getAccountTransactionsStream isnt registered until we
await it's content stream later on. Unfortunately because getNetworkInfo
prevents that form happening it'll be treated as an unhandled exception
which crashes in oclif itself.