Add a development-only feature to print statistics related to notes encryption
This commit adds an optional feature (that can be enabled only during
local development) to print the number of cryptographic operations
performed for note decryption and encryption.

This feature is enabled by building `ironfish-rust-nodejs` through `yarn
build:stats`. After doing that, and after starting `ironfish-cli`
through `yarn start <command>`, it is possible to trigger the printing
of the statistics by sending `SIGUSR1` to the CLI process.

This commit has no impact on release builds of Iron Fish.