fixes db transactions to always have caches
LevelupTransaction is the only type we have currently that implements
IDatabaseTransaction. it implements a caching layer that is flushed to disk when
the transaction commits.

when implementing BlockchainDBTransaction we still need this cache so that data
isn't flushed to disk while the transaction is pending

expands IDatabaseTransaction to include a cache and methods for has, get, put,
add, and del. this allows BlockchainDBTransaction to forward calls to these
methods to the underlying tx, which will always be a LevelupTransaction

removes a couple of type signatures that aren't used for put and add taking
unknown values