implements BlockchainDBTransaction to wrap EVM state changes (#5083)
* wip: implements BlockchainDBTransaction to wrap EVM state changes

the ethereumjs packages implement their own systems for committing changes to
the underlying database

plumbing our own database transaction objects down through several layers will
require significant refactoring to evm caching and database access

BlockchainDBTransaction wraps an IDatabaseTransaction and inserts logic to
use the evm stateManager's checkpoint logic for committing and reverting state
changes

implements transaction and withTransaction methods on the BlockchainDB and
replaces access of the underlying db instance with these transaction methods to
ensure that BlockchainDBTransactions are used

moves stateManager directly onto BlockchainDB to support access in transaction
context

* updates test-evm command

* updates verifier.evm.test

* fixes db transactions to always have caches

LevelupTransaction is the only type we have currently that implements
IDatabaseTransaction. it implements a caching layer that is flushed to disk when
the transaction commits.

when implementing BlockchainDBTransaction we still need this cache so that data
isn't flushed to disk while the transaction is pending

expands IDatabaseTransaction to include a cache and methods for has, get, put,
add, and del. this allows BlockchainDBTransaction to forward calls to these
methods to the underlying tx, which will always be a LevelupTransaction

removes a couple of type signatures that aren't used for put and add taking
unknown values

* fixes test description

removes incorrect 'not'

* changes IDatabaseTransaction cache to generic Map interface

updates typedefs for BufferMap to conform to Map interface