Wallet cleanup should use an iter to load records
In the case that the records to cleanup is very high, we shouldn't load
all of them in memory before cleaning up. We should only load 1 leveldb
memory block at most. This will ensure if there is a bad feedback loop
this list won't permanently grow and continue to use more memory as it
tries to clean up accounts.

This also ensures that the account record deletion is considered a
record so deleting thousands of empty account cleanup records doesn't
run forever.