stops scanState meter on scan completion
when started the Meter class sets an interval to update its underlying arrays.
if the meter is not stopped then this interval will keep running and keep the
meter in the heap and leaks the meter memory.

the WalletScanner creates a new ScanState and a new Meter for each scan. when
the wallet is at the head of the chain this creates a new scan each time a block
is added.

fixes the memory leak by ensuring that the ScanState Meter is stopped when the
scan completes and its interval function is cleared