Optimize `WalletDB.loadNoteHash()` to avoid hitting the database if the note nullifier is known to be absent
This commit adds an in-memory bloom filter to remember which nullifiers
are stored in the database. The bloom filter is used by
`WalletDB.loadNoteHash()` to avoid querying the database if the given
nullifier is known to be not present.

While at it, `WalletDB.getTransactionHashFromNullifier()` was optimized
in the same way.