Changes how EVM is instantiated.
Previously, we needed to check whether the EVM on the blockchain object was instantiated.
There shouldn't be a case where EVM is not instantiated. We should always have an EVM instance.

That change was made because instantiating EVM requires async code because of the VM instantiation.

This change changes that pattern to a synchronous instantiation. The VM instantiation now happens in the open function that is no longer static.
We check whether the object has been instantiated in the sub functions of the EVM class.