refreshes WalletScanner scanningAccounts on disconnect
the WalletScanner uses a hash, 'scanFrom', for each account in its
scanningAccounts list to check whether it should decrypt notes for the account.
if scanFrom is null or equal to the previous block hash then notes are decrypted
when connecting a block.

scanFrom is set to the account head when calling refreshScanningAccounts and set
to null when connecting a block if scanFrom is equal to the previous hash.

however, when blocks are disconnected scanFrom is not currently changed. in the
event of a reorg, decryption is skipped when connecting blocks from the fork if
'scanFrom' is still set to the hash of a disconnected block.

regenerates fixtures for test of this behavior and updates disconnectBlock to
set scanFrom to null for all accounts where it is equal to the disconnected
block hash.