Add the ability for CLI commands to track and close an RPC client
With the addition of the JSON changes, which requires us to await on the start
command in the IronfishCommand, commands that used `this.sdk.connectRpc()` were
hanging because the clients were not being closed. This change allows us to
fetch a client from the IronfishCommand directly, which gives the
IronfishCommand the ability to track the client, and close it when the code is
done being ran/executed.