adds StateManager method for reading at past state roots (#5181)
* adds StateManager wrapper for reading at past state roots

reading the confirmed balance (for any confirmation range) for an account will
require reading from the state trie at a past state root

adds a wrapper function, withStateRoot, to temporarily set the state root of the
stateManager, execute a handler function to read from the stateManager at that
root, and finally restore the root to its current value

defines getAccount and getBalance on the IronfishEvm using withStateRoot

some RPC methods, like eth_getAccount and eth_getBalance, will require this
functionality

* ensures that handler finishes before executing finally in withStateRoot

* updates withStateRoot to use shallowCopy

uses a copy of the stateManager instead of setting the state root on the
blockchainDb's stateManager. using a copy seems safer since it doesn't risk
affecting the state of the blockchainDb's instance.  the ethereumjs monorepo
also uses shallowCopy for this purpose in its rpcs (e.g., getBalance)