make secret optional in multisig secrets
We need to make the secret field optional to accommodate the ledger device usecase. The secret never leaves the device.

We can write a migration where we add a flag that indicates if the secret exists or not.

But we thought of an alternative that doesn't require a migration - store an empty buffer if the secret doesn't exist. When deserializing, if the buffer is all zeros, then you know the secret is undefined.