Feat/ledger dkg cli (#5421)
* updates zondax/ledger-ironfish dependency to dev branch

* updates zondax/ledger-ironfish-js to local dependency

* updates zondax/ledger-js to dev branch on iron-fish fork

* updates ledger.ts for upgrade to @zondax/ledger-js (#5386)

* updates ledger.ts for upgrade to @zondax/ledger-js

removes response handling for error codes, which are no longer included in
responses

implements tryInstruction to handle errors from any app instructions

adds methods to help discriminate between KeyResponse types and checks type of
KeyResponse for each type of key

adds method to recognize Ledger response errors

defines classes for common recoverable Ledger errors: locked device and app not
open. the error code for the app not being open is identified in the docs as a
'technical error', but this is the error code we get when the app isn't open

* adds ledger flag to participant creation (#5387)

* adds ledger flag to participant creation

adds dkgGetIDentity method to ledger util module

adds '--ledger' flag to 'wallet:multisig:participant:create' command

uses 'wallet/multisig/importParticipant' RPC to add the participant identity to
the walletDb

* implements ledger dkg round1 (#5389)

* implements ledger dkg round1

adds dkgRound1 method to Ledger util

fills in logic of performRound1WithLedger

reads identity from node and adds it to list of signer identities if it's
missing

* implements ledger dkg round2 (#5390)

* implements ledger dkg round2

adds dkgRound2 method to Ledger util

implements performRound2WithLedger in round2 CLI command

* implements ledger dkg round3 (#5391)

* implements ledger dkg round3

implements dkgRound3 method on Ledger util class

implements dkgRetrieveKeys on Ledger util class to retrieve all shared multisig
keys

implements dkgGetPublicPackage on Ledger util class to retrieve public key
package

updates round3 command to run dkg round3, retrieve keys, get public key package,
and import account with ledger flag

* includes participant's own gsk share in round3 input

* renames createParticipantWithLedger to getIdentityFromLedger

* fixes Ledger approval output in round3 (#5397)

ask to approve round3 on Ledger, but don't ask for approval/confirmation when
retrieving keys or public package

* adds ledger option for creating signing commitment (#5398)

* adds ledger option for creating signing commitment

implements dkgGetCommitments on Ledger util class

adds '--ledger' flag to 'wallet:multisig:commitment:create'

adds multisig account selector to 'commitment:create', fetches the identity for
the selected account

adds 'hash' method to UnsignedTransaction primitive

* adds ledger flag for creating signature shares (#5399)

implements dkgSign on Ledger util class to obtain a signature share from the
device

adds '--ledger' flag to 'wallet:multisig:signature:create' command

adds selector to command to choose multisig account if one is not provided

adds 'publicKeyRandomness' method to UnsignedTransaction primitive to extract
the randomness from the transaction (needed for signing)

adds 'serialize' napi method to NativeSignatureShare to support constructing the
signature share from its raw parts and then printing it

* improves error message from importParticipant RPC (#5402)

includes name of conflicting identity in error message if the request tries to
import an identity that already exists in the database

* implements ledger multisig backup command (#5400)

* implements ledger multisig backup command

adds a CLI command, 'wallet:multisig:ledger:backup', to create an encrypted
backup of multisig keys from the ironfish dkg ledger app

users can restore the keys to their ledger app if they reinstall the app on
their device or overwrite the multisig keys in the app

* adds cli command to restore ledger multisig backup (#5401)

* adds cli command to restore ledger multisig backup

if the ironfish dkg ledger app is uninstalled, or if it is used to create a
second set of keys, then the multisig keys on the device may be lost

the 'wallet:multisig:ledger:backup' command creates an encrypted key backup that
can be restored to the device

the 'wallet:multisig:ledger:restore' command restores the backed up keys to the
ledger

* Update ironfish-cli/src/commands/wallet/multisig/ledger/restore.ts

Co-authored-by: mat-if <97762857+mat-if@users.noreply.github.com>

---------

Co-authored-by: mat-if <97762857+mat-if@users.noreply.github.com>

* updates command description

adds additional log output asking the user to save the backup and explaining how
to restore the keys

* fixes lint

---------

Co-authored-by: mat-if <97762857+mat-if@users.noreply.github.com>

* feat: DKG Create Command (#5403)

* adds ledger backup to the end of round3 (#5406)

creates an encrypted backup and prints it to the terminal at the end of round3
of dkg

* prompts for new name and retries import in round3 (#5407)

extracts the import/name prompt loop logic from the 'wallet:import' command into
an 'importAccount' util

uses the 'importAccount' util to import the account at the end of round3 so that
import is retried in the case of name collisions

* adds cli command to import multisig account from ledger (#5409)

reads identity, shared multisig keys, and public key package from Ledger device

imports multisig account to wallet and prompts for new name if the provided name
is taken

removes hardcoded account version number from other ledger imports (these
imports aren't from serialized accounts, so the schema version should always be
the current version)

* adds clear signing to ledger signing flows (#5413)

uses 'reviewTransaction' method from '@zondax/ledger-ironfish' to force user to
review decrypted transaction outputs on the ledger device

updates our '@zondax/ledger-ironfish' dependency to use the latest release,
which now includes dkg methods

note: for interactive signing we may only need to review once before creating
the signing commitments

* always display asset id with notes in transaction details

when approving a transaction using a ledger device users are shown the asset id
of each output note on the ledger device

the transaction details they see in the cli will not include the asset id if the
asset is verified (the symbol will be displayed instead). this makes it more
difficult to verify that the transaction you're seeing on the ledger is correct
based on what the cli is showing you

if the asset has a verified symbol, also displays the asset id for the note on
the line below the amount

* Revert "always display asset id with notes in transaction details"

This reverts commit e7f3e46ff68460f57c9f39a417812782a0e9fec7.

* always display asset id with notes in transaction details (#5416)

when approving a transaction using a ledger device users are shown the asset id
of each output note on the ledger device

the transaction details they see in the cli will not include the asset id if the
asset is verified (the symbol will be displayed instead). this makes it more
difficult to verify that the transaction you're seeing on the ledger is correct
based on what the cli is showing you

if the asset has a verified symbol, also displays the asset id for the note on
the line below the amount

* Adding ledger support for dkg:create (#5410)

Also added the ability to reuse an existing identity for the DKG process.
The user can specify the account name separately from the participant name.

* multisig:sign command (#5411)

* fix typo

---------

Co-authored-by: Hugh Cunningham <hugh@iflabs.network>
Co-authored-by: Hugh Cunningham <57735705+hughy@users.noreply.github.com>
Co-authored-by: mat-if <97762857+mat-if@users.noreply.github.com>