Expect to participant in signing
Previously, this command was written in a way where the user performing this command didn't have to be a participant.
This changes that to make sure that the user performing this command is a participant in the multisig wallet.

This is because this command is catering to the most common use case.