adds multisig server to broker dkg messages
copies stratum module from sdk into cli to serve as foundation for multisig
network module

replaces stratum mining messages with dkg messages:
- 'dkg.status': returns the dkg status of the session including minSigners,
  identities, and all public packages
- 'dkg.get_status': requests the status of the dkg session from the server
- 'dkg.identity': sends an identity over the socket
- 'dkg.round1': sends a round1 public package over the socket
- 'dkg.round2': sends a round2 public package over the socket

reuses single message for submitting an identity to the server and broadcasting
it from the server. likewise for round1 public packages and round2 public
packages

adds a cli command, 'wallet:multisig:server', to start a server for a dkg
session. requires maxSigners and minSigners on start to set the parameters for
DKG

updates 'wallet:multisig:dkg:create' with a '--server' flag to connect to a
server. at each step, submits the participant's information to the server and
then waits for server to broadcast packages from other participants. uses
'dkg.get_status' to get the status at the beginning of each step to account for
the possibility of having missed broadcasts