Fix vscode not resolving test types (#5294)
With some recent changes to our jest types, vscode no longer resolved
`describe`, `expect`, etc.

I believe the issue is that vscode's typescript extension relies on
`tsconfig.json`, without giving users the ability to change that. Our default
tsconfig.json excluded our test files, which is what was causing the issue. I
do not know why this wasn't an issue when we used `@types/jest`, but there
seems to be a lot of special-casing around `@types`, so maybe it was a defualt
tsconfig option somewhere that I wasn't able to figure out.

In the mean time, if we flip our usage of tsconfig - make tsconfig.json the one
that includes all tests files, and tsconfig.build.json the one we use for
building the actual deliverable code, we allow vscode to do what it wants to
do, while hopefully still only building the files we want