encrypts multisig server messages
uses xchacha20poly1305 to encrypt all string fields in messages sent to multisig
server

client decrypts string fields in messages received from multisig server

derives the client key from a passphrase and the session ID (uses the bytes of
the session ID, which is a UUID, for the salt and nonce). ensures that any
client in the session can derive the key if they have the passphrase.

adds passphrase flags and prompts to dkg:create and multisig:sign

NOTE: numeric fields, like minSigners, are not currently encrypted

Closes IFL-3013