Adding zondax/ledger-js dependency (#5466)
zondax/ledger-js is a subdependency of @zondax/ledger-ironfish. We rely on the ResponseError type from ledger-js which should ideally be exported from zondax/ledger-ironfish. This way we don't have to explicitly include this subdependency in our package.json.

A todo has been added to remove this dependency when the ResponseError type is exported from ledger-ironfish.