automatically retries messages to multisig broker
adds a server 'ack' message that the server sends to the client upon receiving
and successfully parsing the client message. each 'ack' contains the client's
message id

client sets an interval to retry sending each message. when it receives the
'ack' for its message it clears the interval for that message id