always reviews tx in same connection as commit or sign
clear signing in the Ironfish DKG Ledger App requires that the client review the
transaction with 'reviewTransaction' before generating signing commitments with
'dkgGetCommitments' or creating a signature share with 'dkgSign'

an approved 'reviewTransaction' instruction stores the transaction hash of the
approved transaction in the device memory. however, storage of that hash does
not persist between transport connections. so, if a connection is lost due to an
error, or if a new connection is created for each instruction, then the hash
cannot be accessed to allow generation of commitments or signature shares

modifies 'dkgGetCommitments' and 'dkgSign' in 'LedgerMultiSigner' to take an
unsigned transaction and always call the 'reviewTransaction' instruction in the
same connection as 'dkgGetCommitments' and 'dkgSign'