retries joining session if passphrase incorrect
if a user enters the passphrase for a session by the client is unable to decrypt
the session challenge string, the client emits a 'SessionDecryptionError'. the
'wallet:multisig:dkg:create' and 'wallet:multisig:sign' commands will now retry
the step and prompt the user to re-enter the passphrase

adds an event emitter, 'onClientError', to the multisig client. emitting client
errors as events provides a way for the command logic to define error handlers
to throw/catch the error in the control flow of the command instead of in
response to server data