clears onDisconnected event handler when ending session
the handler that we add to the MultisigClient.onDisconnected event assumes that
the client will try to reconnect and rejoin a session after a disconnect.
however, if the session is ended, the client will not try to reconnect. this
results in the CLI displaying an endless action stating that it is waiting to
connect after the session ends

clears the handler before stopping the client so that the sessionManager won't
wait for messages confirming server reconnection and session rejoin