split very long running test into smaller tests
this test sometimes takes over 4 minutes on CI. by splitting it up, we
normalize the run time across the test shards, creating a more consistent run
time across the CI shards. other upside is that if one of these test fails, it
will be easier to see at a glance which one is failing by the name.