Feature-gate all code in ironfish-rust that requires Sapling parameters
This commit adds a new `transaction-proofs` Cargo feature (enabled by
default) to the `ironfish-rust` crate. When this feature is disabled,
all the structs and methods that create or verify zero-knowledge proofs
are also disabled, hence making the Sapling parameters not needed.

This way, developers have an option to use some of the `ironfish-rust`
features (like transaction decryption) without having to deploy the
Sapling parameters, which are several MBs in size.

When `transaction-proofs` is disabled, all the transaction builder
structs (like `ProposedTransaction`, `SpendBuilder, `OutputBuilder`,
`MintBuilder`, ...) and related methods are unavailable, because those
structs/methods all generate proofs.

This commit also cleans up and reorganizes some code, in particular:

* renamed the `ironfish::transaction::utils` module to
  `ironfish::transaction::verify`, because this module only contains
  methods for verification;
* moved some of the transaction verification methods from
  `ironfish::transaction` to `ironfish::transaction::verify`;
* moved `ProposedTransaction` to a dedicated submodule;