Make the Sapling parameters optional for documentation builds
This is to allow our documentation to build on platforms like docs.rs,
which need to build the documentation for our crates, but don't have
access to the Sapling parameters.

This also slightly improves the experience for developers who use this
crate for the first time: if they use the crate without the Sapling
parameters, they'll now get a helpful warning instead of a panic.