supports migrations on encrypted wallets
adds a template migration , 033-encrypted-wallet-template, to demonstrate how to
write wallet migrations on wallet databases that may be encrypted

now that account data may be encrypted we need to be careful to decrypt the
account before running the migration and to re-encrypt the migrated data before
writing it back to the database

adds EncryptedWalletMigrationError thrown when a migration tries to access an
encrypted account without the wallet passphrase to allow client code (e.g.,
'migrations:start' command) to handle decryption flow

optionally passes wallet passphrase through migrator to an individual migration
to decrypt/encrypt account data in migrations

updates 'start' command to catch EncryptedWalletMigrationError and prompt for
wallet passphrase. passes wallet passphrase through node openDB flow to suppoprt
running migrator with wallet passphrase

NOTE: this example assumes that the schema for the masterKey store does not
change and assumes that the schema for encrypted account data does not change.
if the masterKey schema changes then that change must be addressed in a separate
migration. if the encrypted account schema changes, then that change should be
addressed separately from any change to the decrypted account schema in a
separate migration.