WASM: enable in-browser tests through `wasm-bindgen-test`
The unit tests currently are marked with both `#[test]` and
`#[wasm_bindgen_test]`. This allows tests to run with both `cargo test`
and `wasm-pack test`.

In theory, we could drop `#[test]`, but given that `cargo test` is
faster than `wasm-pack test`, this makes developing easier, so I'm
keeping it for now.