handles Ledger 'invalid dkg status' errors (#5667)
if a user attempts to sign a transaction, but their ledger app doesn't have any
multisig keys persisted, then the signing operation will fail with an
'invalid dkg status' error. this can happen if the user reinstalls the app or if
the user starts creating a new account on the device using bkg, but doesn't
finish

the only ways to recover from this error are to restore an encrypted backup to
the device, or to create a new multisig account using the device. since the
error shows up during signing, we can infer that the user expects to have an
account on the device and should be instructed to restore the backup