Implement spend and mint description builders
These are feature-gated and disabled by default because they do not work
in a web environment. In order to work, users need to patch `yastl` so
that it runs tasks serially rather than in parallel (because there's no
support for threads in a web environment). Also, even after patching,
generating a zero-knowledge proof currently takes an absurd amount of
time in a web browser.